// packages/api-client/src/types/Network/MobileOperator.ts
/**
 * Copyright 2023-2024 MICRORISC s.r.o.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

/**
 * Mobile operator
 */
export interface MobileOperator {
	/// Operator APN
	apn: string;
	/// Operator ID
	id?: number;
	/// Operator name
	name: string;
	/// Operator password
	password?: string;
	/// Operator username
	username?: string;
}